Entity registry v1 (no components!)
