+++ Ranitas/Core/ECS/EntityRegistry.cs
@@ -1,0 +1,46 @@
+﻿using System;
+using System.Diagnostics;
+
+namespace Ranitas.Core.ECS
+{
+    public class EntityRegistry
+    {
+        public EntityRegistry(int maxEntities)
+        {
+            Debug.Assert(maxEntities <= Entity.MaxIndex, "Can't store this many entities!");
+            mEntities = new Entity[maxEntities + 1];
+            mEntities[0] = new Entity(0, uint.MaxValue); //Index 0 is never assigned, and NullEntity is invalid.
+            mNext = 1;
+            for (uint i = 1; i <= maxEntities; ++i)
+            {
+                mEntities[i] = new Entity(i + 1, 0);    //Point to the next entity available
+            }
+        }
+
+        private uint mNext;
+        private Entity[] mEntities;
+
+        public Entity Create()
+        {
+            uint createdIndex = mNext;
+            Debug.Assert(createdIndex < mEntities.Length, "Entities have run out!");
+            Entity inPlaceEntity = mEntities[createdIndex];
+            mNext = inPlaceEntity.Index;
+            Entity createdEntity = new Entity(createdIndex, inPlaceEntity.Version);
+            mEntities[createdIndex] = createdEntity;
+            return createdEntity;
+        }
+
+        public void Destroy(Entity entity)
+        {
+            Debug.Assert(IsValid(entity), "Can't destroy invalid entity.");
+            mEntities[entity.Index] = new Entity(mNext, entity.Version + 1);
+            mNext = entity.Index;
+        }
+
+        public bool IsValid(Entity entity)
+        {
+            return mEntities[entity.Index] == entity;
+        }
+    }
+}
